02-docker/requirements.txt
google-cloud-firestore==2.19.0
click==8.0.1
flask>=2.2.5
gunicorn>=22.0.0
itsdangerous==2.0.1
jinja2>=3.1.5
jmespath==0.10.0
MarkupSafe>=2.1.1
python-dateutil==2.8.1
six==1.16.0
urllib3>=1.26.19
Werkzeug>=3.0.6